chore: move services tests into dedicated folder
